--- src/commonMain/kotlin/io/konform/validation/kotlin/Path.kt
@@ -1,5 +1,0 @@
-package io.konform.validation.kotlin
-
-import kotlin.reflect.KFunction1
-import kotlin.reflect.KProperty1
-
